+++ client/src/components/ChoosingColor/ChoosingColor.style.js
