--- client/src/components/atoms/Hamburger.style.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Hamburger = styled.img`
-  width: 30px;
-  height: 30px;
-  cursor: pointer;
-  z-index: 99;
-`;
-
-export default Hamburger;
