--- client/src/components/atoms/Logo.style.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-const Logo = styled.img`
-  width: 35px;
-  height: 40px;
-`;
-
-export default Logo;
