+++ client/src/layouts/MainLayout/MainLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import Navigation from '@/components/organisms/Navigation/Navigation.jsx';
+import Navigation from '@/components/Navigation/Navigation.jsx';
 import MainLayoutContainer from './MainLayout.style';
 
 function Layout() {
