+++ client/src/views/Register/Register.jsx
@@ -3,10 +3,10 @@ import { useNavigate } from 'react-router-dom';
 import { useForm } from 'react-hook-form';
 import { useDispatch } from 'react-redux';
 import { setCredentials } from '@/features/auth/authSlice';
-import AuthForm from '@/components/organisms/AuthForm/AuthForm.style';
-import AuthTitle from '@/components/atoms/AuthTitle/AuthTitle.style';
-import AuthTextField from '@/components/molecules/AuthTextField/AuthTextField.style';
-import AuthSubmitButton from '@/components/atoms/AuthButton/AuthButton.style';
+import AuthForm from '@/components/Auth/AuthForm/AuthForm.style';
+import AuthTitle from '@/components/Auth/AuthTitle/AuthTitle.style';
+import AuthTextField from '@/components/Auth/AuthTextField/AuthTextField.style';
+import AuthSubmitButton from '@/components/Auth/AuthButton/AuthButton.style';
 import mailIcon from '@/assets/images/icons/mail.svg';
 import passwordIcon from '@/assets/images/icons/lock.svg';
 import { useLoginMutation, useCreateNewUserMutation } from '@/features/auth/authApiSlice';
