+++ server/config/allowedOrigins.js
@@ -1,3 +1,3 @@
-const allowedOrigins = ['https://financial-checker.onrender.com'];
+const allowedOrigins = [process.env.FRONT_END_ADDRESS];
 
 module.exports = allowedOrigins;
